refactor(tests): reorganize async tests
