+++ crates/misc/component-async-tests/Cargo.toml
@@ -7,17 +7,24 @@ edition.workspace = true
 rust-version.workspace = true
 publish = false
 
-[dev-dependencies]
+[dependencies]
 anyhow = { workspace = true }
 flate2 = "1.0.30"
 futures = { workspace = true }
 pretty_env_logger = { workspace = true }
 tempfile = { workspace = true }
-test-programs-artifacts = { workspace = true }
-tokio = { workspace = true, features = ["fs", "process", "macros", "rt-multi-thread", "time"] }
+tokio = { workspace = true, features = [
+  "fs",
+  "process",
+  "macros",
+  "rt-multi-thread",
+  "time",
+] }
 wasi-http-draft = { path = "http" }
 wasm-compose = { workspace = true }
 wasmparser = { workspace = true }
 wasmtime = { workspace = true, features = ["component-model-async"] }
 wasmtime-wasi = { workspace = true }
 
+[dev-dependencies]
+test-programs-artifacts = { workspace = true }
