+++ crates/misc/component-async-tests/src/proxy.rs
@@ -1,0 +1,42 @@
+
+use anyhow::{anyhow, Result};
+use wasi_http_draft::wasi::http::types::{ErrorCode, Request, Response};
+use wasi_http_draft::WasiHttpView;
+use wasmtime::component::{
+    Accessor, Resource, ResourceTable,
+};
+use wasmtime_wasi::IoView;
+
+
+pub mod bindings {
+    wasmtime::component::bindgen!({
+        path: "wit",
+        world: "wasi:http/proxy",
+        concurrent_imports: true,
+        concurrent_exports: true,
+        async: {
+            only_imports: [
+                "wasi:http/types@0.3.0-draft#[static]body.finish",
+                "wasi:http/handler@0.3.0-draft#handle",
+            ]
+        },
+        with: {
+            "wasi:http/types": wasi_http_draft::wasi::http::types,
+        }
+    });
+}
+
+impl WasiHttpView for super::Ctx {
+    type Data = super::Ctx;
+
+    fn table(&mut self) -> &mut ResourceTable {
+        &mut self.table
+    }
+
+    async fn send_request(
+        _accessor: &mut Accessor<Self::Data>,
+        _request: Resource<Request>,
+    ) -> wasmtime::Result<Result<Resource<Response>, ErrorCode>> {
+        Err(anyhow!("no outbound request handler available"))
+    }
+}
