+++ crates/misc/component-async-tests/src/round_trip_direct.rs
@@ -1,0 +1,25 @@
+use std::time::Duration;
+
+use wasmtime::component::Accessor;
+
+use super::Ctx;
+
+pub mod bindings {
+    wasmtime::component::bindgen!({
+        trappable_imports: true,
+        path: "wit",
+        world: "round-trip-direct",
+        concurrent_imports: true,
+        concurrent_exports: true,
+        async: true,
+    });
+}
+
+impl bindings::RoundTripDirectImports for super::Ctx {
+    type Data = Ctx;
+
+    async fn foo(_: &mut Accessor<Self>, s: String) -> wasmtime::Result<String> {
+        tokio::time::sleep(Duration::from_millis(10)).await;
+        Ok(format!("{s} - entered host - exited host"))
+    }
+}
