+++ crates/misc/component-async-tests/src/transmit.rs
@@ -1,0 +1,11 @@
+
+
+
+pub mod bindings {
+    wasmtime::component::bindgen!({
+        path: "wit",
+        world: "transmit-callee",
+        concurrent_exports: true,
+        async: true,
+    });
+}
